New toolbar on bottom of posts screen.
The new toolbar is thinner and looks nicer than the stock UIToolbar we used
before. Since it's thinner, there's no longer any need to have different
metrics for landscape on the iPhone.

Unfortunately UIToolbar didn't want to cooperate with vertically aligning
items on a shrunken toolbar, so we're rolling our own view for it.