Load correct page after replying.
Before we just reloaded the current page, which fails when, say, replying
to the 40th post on the last page, or replying to a post not on the last
page of a thread.

Fortunately, the SA Forums will take a post's ID at a certain URL and
redirect you to the correct thread and page.

We were also doing URL query parsing in two different places, so that's
factored out into a category on NSURL.

Finally, we'll start moving to a one-callback API for AwfulHTTPClient.
Requiring separate success and error blocks makes us repeat easily
forgotten code (e.g. hiding a HUD).