Fix shadow under posts view while refreshing.
To show the refresh control, we change the web view's scroll view's
content inset. This caused the web view to draw a shadow beyond the
refresh control, where previously it had drawn the shadow under the
refresh control itself.

The web view does this by adding an image view to its scroll view, so we
need to get rid of that. By adding the shadow to the refresh control
itself, we get our nice shadow where we want it.