Fix crash refreshing bookmarks after deletion.
If you delete a bunch of bookmarks outside of the app, the next refresh of
bookmarks in the app would crash. It would get a message to update an
object at an index path that was out of bounds, or should have been when
considering all the deletions. Not sure if this is our bug, but here's a
workaround.