Fix iPad crash showing jump to page after reply.
Sending a reply causes a page to (re)load. Some forgotten code used to try
to hide the jump to page view whenever a new page was loaded. It did this
by tapping the button to show/hide the view. Except popovers don't like
being shown in views that don't have a window, so it would throw an
exception when we accidentally tried to show it while the reply view was
visible.