Don't jump around so much showing read posts.
The goal when showing read posts is not to scroll anywhere. That's not
possible when there's less than a screenful of unread posts, as they'll go
to the bottom after loading read posts (pushing us over a screenful).
Moreover it's easy to accidentally trigger a pull to refresh. Now we just
scroll to the bottom if staying where we are isn't possible.