Fix parsing whether a thread has been seen.
The seen class is only on the row if the "Highlight threads I've seen in a
different color" option is enabled on the user's Forums account. The
div we look for now doesn't depend on an option.

I think we used to use the div and I switched it to the "seen" class on the
row. I switched it for two reasons: I thought we did it the old way because
we didn't used to search for classes using XPath; and I didn't realize the
"seen" class was the manifestation of an option.