Fix next page button refreshing current page.
If you refresh the last page and it's no longer the last page, the UI would
update properly but tapping the next page button wouldn't work. It would
only reload the current page, after which the next page button worked.

The problem is we were scraping the updated page count correctly, but we
would inadvertently overwrite it immediately thereafter, when marking the
posts as read. The total number of replies value was out of whack, so we'd
"fix" it, which would decrease the number of pages value to match the
number of replies.

The underlying problem here, maintaining both a number of pages attribute
and a total number of replies attribute, remains to be fixed later.