Fix saving parsed user profiles.
By not setting the user ID we were creating then updating a User instance with
a nil user ID. This appeared in the interface when viewing a profile that had
e.g. an AIM username, then viewing a profile that did not. The first user's AIM
username would erroneously appear under the second user.