Update setting immediately from choices view.
This is when choosing e.g. a YOSPOS theme. Before it wouldn't take effect until
leaving the list of choices. This looks dumb on iPad, when you have the choice
screen open with a thread in behind.

The settings view now reloads on appear so it's up-to-date after the choice is
made.