Thread cell shows title right up to unread count.
The hardcoded 70 value seemed to be enough to show a five-digit unread count.
It's smarter not to hardcode this in case there's more unread posts in someone's
thread. Additionally, this shows more of a thread's title in the much more
frequent case that there are fewer than five digits' unread posts.