Remove "Keep Sidebar Open" setting.
Just show it in landscape, hide it in portrait.

I could not get UISplitViewController to hide/show the master view immediately
on changing the setting, and I could not get the setting to take effect
consistently on both iOS 5 and iOS 6 without quitting and reopening the app.
This seems like the most sane setting, so let's just leave it.