Nicer profile view based on @ridilla's mockup.
Mockup looked great. Now shown on iPad as a modal view, with the two sections
stacked horizontally. This fixes #156.

Profile view can bounce when it scrolls, so it doesn't feel weird anymore.

And there's no more gap at the top of a profile view for users with no avatar or custom title.