Update model to be more like SA.
Threads have no concept of being "locked"; they're either open or closed.
Closed threads are shown greyed out, unless they are sticky.

Posts are editable only by their author, and only if the thread isn't archived.

Attachments and "edited by" isn't part of the HTML of a post in JSON.

AwfulThread doesn't need a "seen" property in the model, as it can be
implemented off of the "totalUnreadReplies" property.

AwfulPost doesn't need a beenSeen property, as it can be implemented via
-[AwfulThread totalUnreadReplies]