Fix NSURLErrorDomain error -999.
Since networking isn't always up when we refresh a list in -viewWillAppear:, we
also refresh when network reachability changes. If this happens quick enough,
we could end up refreshing a second time before the first refresh is done. Now
we only refresh on reachability changes if we're not already refreshing.