Fix crash when migrating data store.
The favorites migration gets and sets the AwfulSettings favoriteForums array as
it does its work. Unfortunately, as the Core Data stuff gets lazily
initialized, this causes e.g. the favorites controller to try fetching
AwfulForums in the middle of the migration. When this happens, the Core Data
stuff isn't initialized, and we get a crash.