Better fix for wrong thread cell data.
Earlier fix was 5622a150d6ed2dce37d91404351eff4f3c85ad9a.

Reloading the row messes with other ongoing animations. Here we pass in the
object if we already have it, so the -configureCell:... method doesn't grab the
wrong object. If we don't already have the object, they can figure it out.