Hackily respect 'Show bookmark category "star buttons" next to threads' setting.
The idea is to hide the colored star buttons for bookmarks, which we want to
interpret as "don't show bookmark colors". Unfortunately, the Forums implements
this using CSS, which we don't currently parse. So instead we assume that, if
all the bookmarked threads are the default color, we shouldn't use colors.

Basically I want blue badges because I don't use colors, but people who do use
colors need to see the actual color from the Forums.