Clean up "view user's posts" model and API changes.
Moved the model changes to a new model version. Also switched to a
"singleUserIndex" for posts, instead of simply a boolean, as we can't tell a
post's overall thread index from a page of a single user's posts.

Changed relevant API to refer to a "singleUserID", and removed the variants
that don't take a singleUserID so there's just one call to make.

The JumpToPageControllerDelegate doesn't need a singleUserID property; it's up
to implementers of the protocol to decide what "last page" means.