Fix crash loading a new page of posts after scrolling.
We cache post updates while scrolling the posts view so that scrolling doesn't
get all jittery. Unfortunately, by not cleaning out this cache, it was possible
to run the cached updates after we'd moved on to the next page. This can crash
when, for example, a cached update would reload a post index beyond the next
page's number of posts.

The fix is to simply delete the update cache if we're loading a new page.