Fix and simplify two-line title in navigation bar.
Since moving the compose reply button up to the navigation bar, there's always
a left item and right item whenever the two-line title is shown. So it makes
sense to center the text again, and that means we can just let UINavigationBar
do its thing. Happily, that fixes reported issues with title layout when
rotating an iPhone while a YouTube video is fullscreen.