Tweak "show sidebar" gesture on iPad.
A pan gesture recognizer has much nicer interaction with scrollviews' gesture
recognizers. And it easily allows for a "peek" gesture: swipe right to reveal
the sidebar but don't let go; check the threads, then swipe back left to hide
the sidebar once more.

In the meantime, the recognizers are named after what they do, and the two that
hide the sidebar now share an action method.