Remove first stylesheet load detection from posts view.
By far the most common of the many reported WebKit crashes has to do with
finishing loading a resource. The first stylesheet load detection in the posts
view used an onerror handler on an <img> element, which probably gets triggered
after the resource finishes loading. It seems prudent to remove anything even
remotely tricky when loading a resource, so this is gone.