Goodbye mustache.js, hello GRMustache.
The endless WebKit crashes continue, and I'm down to my last idea: move as much
as possible out of the UIWebView. Rendering mustache templates in Objective-C
is the first move. As an added benefit, it cleans up the build process a bit.

This commit also introduces the concept of a "view model" to Awful. These are
model objects for a Mustache template. They're handy for compound properties
(those derived from several properties on the underlying model object), default
values, and formatters. There's one for posts and another for profiles.