Fix forum tree controller's hidden forum set creation and management.
Rows are inserted out-of-order. We need to compute the hidden forum set for the
entire section once it's inserted in order to figure out where inserted rows
are visible.

This issue arose consistently on login, after the forum tree controller is
initialized, when all the forums get inserted as a batch.