Fix state-restored inline images in compose views.
NSTextAttachment doesn't store its bounds when serialized, so we move bounds
calculation into a subclass (which makes more sense anyway).

The other challenge was the image property. NSTextAttachment serializes image
data to a file, and doesn't restore the image property when deserializing.
Since our Imgur uploader wants a UIImage, we reify the image property in the
subclass.

Fixes #279.