Present icon action sheet using WYPopoverController.
WYPopoverController recently gained a "present as dialog" method that just
shows it in the center of the screen with no arrow. This properly handles
non-portrait orientations on iPhone (we were getting a crash with auto layout
spanning rotational bounds transforms).

Finally, always add the background toolbar so we don't get a white background
on iPad.