Rewrite forum tree as a UITableViewDataSource.
A la the fetched results controller data source. This one provides for the
favorites section with a -topDataSource property, freeing the forums list
controller to deal exclusively with the favorites section. Moreover, it's
simpler, which should reduce crashes and bugs and the subsequent inability to
fix them.

Other opportune changes:

* AwfulForumsListController no longer implements
  -{de,en}codeRestorableStateWithCoder: as there is no longer anything to do.
* AwfulForumsListController stops observing the reachability changed
  notification, preventing possible multiple or spurious refresh attempts.
* DRY up a bit of AwfulForumsListController.