Fix bookmarked threads' sorting on iPhone.
We were pausing the thread data source when the view disappeared, then
unpausing on reappear. This is good because table views really hate being
updated when they're off-screen. Unfortunately the data source was turning
around and pausing its backing fetched results controller, so updating a
thread's unread count wouldn't get caught by the data source.

Now, when a thread data source is paused, it simply doesn't tell the table view
anything. It maintains the fetched results controller's change tracking so we
get an up-to-date view when reloading the table view on reappear.