Don't set cache name for forum tree data source.
We were seeing a crash in 2.0.3. An example crash is
http://crashes.to/s/b201e295bac. The crashing method is
-[NSFetchedResultsController(PrivateMethods) _preprocessUpdatedObjects:insertsInfo:deletesInfo:updatesInfo:sectionsWithDeletes:newSectionNames:treatAsRefreshes:].

Some searching turned up http://stackoverflow.com/a/14843264, which mentions a
crash in that method, and finds that an undetected invalid cache is the
culprit. Since we only started setting the cache name in 2.0.3, and turning off
the cache is no big loss, it seems worth trying.