Attempt to fix broken "sort unread threads first" setting.
Unbeknownst to me, if you make certain updates in another managed object
context and use -mergeChangesFromContextDidSaveNotification: to bring the
changes into the main context, a fetched results controller may not react if
the updated objects were faults. To work around this issue, we fire all faults
before merging in the changes.

I have a hard time reproducing the bookmark sorting bug, so I'm not sure this
fixes it. But it seems like a thing we need to do regardless.