Fix state restoration for Awful Browser.
We hadn't set a restorationClass and the app delegate couldn't find it, so we'd
restore the navigation controller but not the actual browser view controller,
leaving people with a black screen and an itemless navigation bar.

In fixing this (by setting ourselves as the restoration class), we're now
saving the restoring the UIWebView's state.