Stop using abbreviated forum names.
Once upon a time, these were used for the back button in navigation items. Now
that all of our back buttons are empty, boring chevrons, there's no need for
the abbreviations.

Their only extant use was in setting the title of a forum's screen, which is
overwritten as soon as the forum refreshes. This was ugly and pointless, and
its removal led to the discovery that nothing uses the abbreviations anymore.