Bring sanity to "mark as read up to post" aftermath.
Instead of re-rendering the whole posts screen whenever marking a thread as
read up to a post, don't. Just update the class attribute of the post elements.

Instead of re-rendering whenever the thread's seenPosts attribute changes, say
due to an automatic bookmarks refresh, don't. Only make visible changes in
response to "mark as read up to post" action.

Instead of showing a pointless HUD after marking a post as read, don't. Let the
change in post backgrounds speak for success.