Show embedded images when previewing posts.
We use yet another custom URL protocol to get the images into a UIWebView. The
only other bit of trickery is that the Forums refuses to parse [img] BBcode
unless the URL resembles an HTTP URL. To work around this, we prefix our custom
URLs with http:// then remove that prefix after fetching the preview.