Rewrite thread tag picker.
The new picker is simpler as it just takes an array of image names (plus an
optional array of secondary image names). The delegate no longer needs to
supply images.

In addition, the secondary tags are more nicely drawn in the picker. The
checkmark seemed superfluous with the filled-in background (plus it obscured
the title text). And the title text has more room to breathe.