Load WebViewJavascriptBridge JavaScript immediately in posts view, PM view, and profile view.
By default, WebViewJavascriptBridge will inject its code after all requests
finish. When images, videos, etc. are slow to load on a page, this causes an
annoying delay until tapping the ... button or a username starts to function.
Loading the bridge's script early solves that problem, and loading it twice
isn't a problem.