Remove superfluous line break when scraping custom titles.
It's there on posts, but not in profiles, so I'm assuming it's something used for styling a page of posts.

Where it was noticeable was in viewing someone's profile. The custom title
would be rendered with the line break at first, then when the profile info
would come in the line break would magically disappear. The change in header
size was painfully obvious once you noticed it.