Switch to (horizontal, UIKit) tab bar controller on all devices.
This is how it was in Awful 1, and to be honest it's the superior interface.

The PM and Leper's Colony view controllers needed more helpful tab bar item titles, and the sidebar on iPad can shrink down now as there's no horizontal space needed for a tab bar.

Bump the interface version stored when preserving state, as they're no longer compatible (the class of the decoded object for the "root" identifier doesn't match). And while we're here, formalize and describe the interface versions.

Don't try to restore old, incompatible state. This means we can get rid of the otherwise pointless AwfulExpandingSplitViewController class, because we're not restoring state that involves it.