Fix "Unbalanced calls to begin/end appearance transitions" on launch.
This comes up when we attempt to show the split view's primary view controller split view too soon. The fix happens in two places:

1. Show the primary view controller from the empty view controller's -viewDidAppear: method.
2. Don't cause the primary view controller to appear when setting the split view controller's preferred display mode.

As a bonus, when the "hide sidebar in landscape" is turned on, the sidebar is no longer inappropriately visible after restoring from state with posts visible.