Switch back to view controller-controlled status bar.
Image previewer wants to hide/show the status bar, so back we go. Only real trick here is that UISplitViewController doesn't provide any hooks for dealing with the status bar (it steadfastly returns nil for the childViewControllerForStatusBar* methods and offers no alternatives). We could handle it in a category, but let's play it safe and aim for a subclass.