Periodically check for and download new smilies.
The new operations are:
* SmilieCleanUpDuplicateDataOperation, which runs immediately after bringing up a smilie data store, before doing anything else. The rest of the framework assumes there are no duplicate smilies.
* SmilieScrapeAndInsertNewSmiliesOperation, intended to be run periodically to look for and download any new smilies.
* SmilieDownloadMissingImageDataOperation, intended to be run on every launch. Probably makes sense to run after a scrape-and-insert operation.

In order to manage those operations, and to simplify framework use, a new SmilieKeyboard class is here. It wraps up a data store, a keyboard view, and an operation queue to manage the new operations.

In addition, the SmilieFetchedDataSource will no longer return smilies that have no image data.

Bonus:
* New TestDataStore class that uses an in-memory read-write store.
* Tidy up smilie tests imports, fixtures into Helpers.h.