Stop forcing keyboard height and remove related workaround.
I'd initially tried to force the keyboard height as part of attempting to avoid all "failure to simultaneously fulfill constraints" errors while the keyboard was loading. Now that I've given up on that goal, it's easier just to let the default sizing take effect. The flow layout weirdness has also disappeared, possibly coincidentally, so no need to work around it.