Redo forums list in Swift, using composable data sources and a metadata entity.
I've been wanting to do the data sources like this for awhile, so I'm happy to get it going. The plan is to eventually switch all the table views over to the new system.

The forum tree is once again done entirely in Core Data. With a bit of prodding, the fetched results controller works great. The ForumMetadata attributes visibleInForumList and showsChildrenInForumList are used to keep track of which forums to show; two attributes are used so that e.g. if one expands every subforum in Games, then collapses Games, when one re-expands Games all the subforums are already expanded. The only downside is that, by fetching ForumMetadata, the fetched results controller don't pick up changes in forum names. Instead we need to do that ourselves.

I suspect that the AwfulForumTreeDataSource (i.e. the AwfulSettings-based one) had some extant issues so I'm not sad to see it go.

Favorite forums have also moved to ForumMetadata and are no longer stored in AwfulSettings. A one-time migration takes place after refreshing the forum list, which will probably happen quite quickly as the model changes will delete the cache.

There was some visual weirdness with rearranging the favorites rows, see comments in ForumListViewController.swift for more info.

The forums list has a refresh control now. I also switched the forums list to use the grouped style of table view as a superior way to stop the section headers from floating. Not sure if I like the all-caps headers though.

Favorite stars now disappear when editing to keep those cells a bit less cluttered. And the last row of each section now gets its separator hidden so there's no goofy little chunk of grey above the section headers.