Move and Swiften data store.
Since we're changing the model and skipping migration, I took the opportunity to move the data store class over to Swift and change the location of the store itself.

A directory is easier to keep track of than whatever detritus SQLite decides to leave behind, so let's just take over a whole directory. This greatly simplifies cleanup.

Speaking of cleanup, instead of moving old data stores into new positions, they're simply deleted. Migration would've failed anyway.

Now that user data (expanded and favorite forums) has returned to the Core Data store, it's no longer appropriate to leave it in the Caches directory. I've moved it to the Application Support folder and excluded it from backup.

Bonus:
* Document and property-ify NSFileManager category methods.
* Move the DataStore and NSManagedObjectContext category into the Models group, out of the unhelpfully-named "Data". The remaining files are set for obsolescence so I'm leaving them for now.