Introduce and use AwfulObjectKey for inserting/updating managed objects.
This does us a couple favors:
1. Object keys make sure that we fully describe objects like User that might (temporarily) only have one of their two primary keys.
2. Object keys conform to NSCoding, making them handier for state preservation and restoration, and suitable for use as dictionary keys.

Also, thanks to AwfulObjectKey, we can generalize upserting from collections of keys and remove that responsibility from other code. This considerably simplifies some of the scraping code.

In addition, there was a bug preserving the state of a PM compose screen for a forwarded message, and there was a bug creating a newly-posted Thread object on the wrong context. They are fixed.