Update Thread.totalReplies when setting its numberOfPages.
There's a mismatch in the info we scrape from thread lists versus a page of posts. The thread list will tell us how many replies there are, but not the number of pages. And the page of posts will tell us how many pages there are, but not the number of replies. Obviously we can infer one from the other, since we set 40 posts per page, but we weren't doing that on a page of posts.

This could manifest itself most obviously when the "Sort unread threads first" setting is enabled. A thread might shoot up to the top of your bookmarks, despite showing 0 unread replies. With this change, we'll probably show the wrong number of unread replies, but it won't be 0.