Swiften MessageListViewController.
As it's the last one using AwfulFetchedResultsControllerDataSource, that can go too.

I finally just moved a "visible" property up to AwfulViewController and AwfulTableViewController, and did away with the little ad hoc ones throughout.

The NewMessageChecker is de-Awful-prefixed and has a simpler API that should hopefully do a better job of keeping the unread message count up-to-date.