Add UIColor-returning subscript to AwfulTheme.
This was a bit tricky to handle with the Objective-C bridging. I ended up moving the objectForKeyedSubscript: implementation over to the extension implemented in Swift so that the UIColor overload could successfully call it. I also had to rename the UIColor-overloaded subscript with an @objc attribute so that there aren't two definitions of what looks to the compiler/runtime like the same method.