Fix centering the image in the image previewer and add double-tap gesture.
Previously the image wasn't centered immediately after loading, though zooming would quickly fix that.

Naturally a full rewrite of the image previewer was completely necessary.

Or not. Maybe I'm just in a "boo IB" downslope.

And we're using a new strategy for centering the contents of scroll view: contentInsets! Who knew?