Unpod HTMLReader.
CocoaPods doesn't do so well with framework targets and app targets in the same project, or at least I haven't found a good way around ending up with duplicate symbols. Previously I was abusing the fact that Smilies.framework ends up linked everywhere we needed HTMLReader. That solution, which was never ideal, falls apart when adding a new framework that also has HTMLReader as a dependency but does not itself depend on Smilies.framework.

Instead we'll check out HTMLReader ourselves and link/embed it as necessary with/in each target. Carthage isn't doing anything here except checking out HTMLReader for us; I figured that was the more sensible approach. Adding a submodule prevents Awful from being buildable immediately after checkout. And manually dragging over bits and pieces of HTMLReader didn't seem wise.

In the process, somewhat deal with the removal of HTMLTextNode from the public headers. We need to finish dealing with this sometime later, as we're still assuming the class exists and fills a certain role.

This also means that the ScrapingTests, SmiliesTests, and SmilieExtractor targets no longer need the Pods header directory manually added to their header search paths.

PSMenuItem also needed an update to fix a compile error.