Extract AwfulForumsClient into AwfulCore.
AwfulCore is now no longer marked as using the extension-safe API because AFNetworking is not marked extension-safe.

The only real annoyance here is testing the scrapers, as they're basically private API and exposing them just for testing seems pointless. Fortunately just a HEADER_SEARCH_PATHS build setting and a bridging header are all we need.

Closes #309.