Enswiften and tidy themes.
AwfulTheme is now Theme. AwfulThemeLoader is gone, subsumed into a lil global variable in Themes.swift.

Theme has subscript overloads for UIColor? and String?, obviating a lot of casting.

The themed UIViewController stuff is now in AwfulViewController.h, because it looks a lot nicer in subclasses to #import "AwfulViewController.h".

Themes.plist can now have sections, which make most themes look much tidier. Any item that's a dictionary works as a section, and the name of the section doesn't matter. And sections can have sections, if that's helpful for some reason.

Some unused keys are now gone from Themes.plist. And keys for collectionView<whatever> are now tagPicker<whatever> because that's the only place they were used.

Oh, and I switched the default toolbar color to black.