Don't crash when deleting favourite forum.
The underlying problem is that we were assuming that the favourite forums data source would get its content change notification before the forum tree data source. Since this was, entirely by accident, the case for so long, everything seemed to work.

When the notifications aren't delivered in this order, we have two problems:

1. The CompoundDataSource's section count cache (startingGlobalSections) is temporarily wrong, leading it to query the wrong data source for an index path that was just deleted.
2. After refreshing the row in the tree, the table view would note that the section count had changed (as the last favourite had disappeared) without a corresponding deleteSection call.

The fix to problem 1 is to simply remove the cache. If it turns out that it runs too slow (it doesn't) we can put it back in a more thoughtful way.

The fix to problem 2 is to handle the favourite deletion as a user-driven change, allowing the forum tree's content change notification to do what it needs to do with accurate row and section counts all around.