Update icons.
We had been reusing the just-their-posts icon for the show-in-thread action, but now we have a dedicated show-in-thread icon so let's use that.

We also have a custom compose icon that looks like the system's but fits in better with the other icons in the app, so we're using that too.