Prevent PM compose screen menu tree and post reply screen menu tree from fighting.
If you have both a draft PM and a draft post hanging around, they both have a CompositionMenuTree responding to the menuDidHide notification, and chaos ensues.

Instead, let's assume that we can muck with the menu only when the associated text view is in the view hierarchy, which is currently a reasonable proxy for "should control of the menu".