Fix massive zoom-out on posts view, messages view, and profile view on iPhones with iOS 9 beta 4.
Not sure if this is an intentional alteration of UIWebView's behaviour or just a quirk of the beta. I don't believe the change will break anything with the old behaviour, though, so we're ok to just leave it in.

This works fine for landscape too, despite device-width meaning "the physical screen width" (as opposed to the width in the current interface orientation), thanks to the initial-scale=1.