Extend gif handling to additional hosts
Now will stop gif autoplaying for images hosted at Kinja/Gawker, giphy,
and gfycat in addition to imgur.

The one problem I’ve found is that the only poster images I could find
for giphy are scaled to 200px wide, so the size changes when you play
and pause, which is not ideal.