Fix crash loading short/empty rap sheets.
This was a stupidity twofer. First, it turns out a nil array of punishments is actually a possible value for when there are no punishments on the requested page, so that's a bad time to call fatalError. Second, the "load more when I scroll to the bottom" detection is hilariously shoddy here and was immediately loading page 2 when page 1 was pretty short. Combine the two and boom.