Fix up jumping to posts
In particular, coalesced tweet loading so that Awful now goes through the page and finds all twitter links to transform into embed divs, and only then does it invoke widgets.js. This means that only one "tweets loaded" event is fired, and we can key off that to know when to reassert our scroll position.

Additionally, moved deleting the post or fractional offset to scroll to until the next loadPage(), to duck the refreshing cached posts blowing away the scroll position so no one could jump to quoted posts.