Fix win1252 characters outside ASCII turning into mojibake when posting.
For example, ü is in win1252, so when we submit a form we should neither turn it into an HTML entity nor percent-escape it. Unfortunately an update to AFNetworking switched to using the Foundation percent-encoding methods that hardcode UTF-8, which resulted in e.g. ü getting percent-encoded into its UTF-8 bytes.

Furthermore, AFHTTPRequestSerializer ignores its stringEncoding property when percent-escaping. That's probably because (Core)Foundation deprecated all percent-encoding methods that take a string encoding as a parameter. Which is ok, because you almost never want to use anything other than UTF-8. Unfortunately, we actually do. If/when CoreFoundation removes the deprecated functions, I guess we do our own percent-escaping?

Finally, AFNetworking's query string-generating bits are all private, so we basically have to reimplement the subset we need.