Turning the clipboard URL setting on will immediately check
A common use pattern would be that someone has turned the feature off and forgotten. Later they could copy a URL enter the app, notice it does nothing, and go to settings to enable it. Normally, nothing would happen until they left and reopened Awful, which isn’t a great experience. Now, Awful will immediately check the clipboard when the setting is turned on.