Possible fix for SFSafariViewController not working?
I don’t understand why this works, but before I applied it I was getting the problem (where SFSafariViewController just sits on a blank page and doesn’t load, with the Done button misaligned) quite reliably. Now I don’t have any issue, so perhaps this is the cause. But maybe there is a better solution.