Fix Parent Forum button atop posts page.
Had a few things going on here. First, making the awful:// URL was failing spectacularly (or at least appeared to be doing so in the debugger), and I figured we could be a little safer.

Next, the URL routing was failing to find the Forums list because UIViewController.firstDescendantOfType wasn't particularly thorough in its search. I recently wrote some very similar code in the app delegate to propagate theme changes to all view controllers, and figured there's no need to repeat ourselves. The mess of AnySequence and angle brackets everywhere is a bit of an eyesore but I'm happy with the result.

Fixes #667.