Delete URL+Awful.swift in favor of AwfulRoute and tidy up nearby code.
UIPasteboard.awful_URL is now coercedURL. awful_URL is a relic from an Objective-C category but the name is confusing: it's not really an "Awful URL", just an Awful extension to UIPasteboard that gets/sets a URL.

UIPasteboard.coercedURL now uses the hasStrings and hasURLs properties when available.

Rename Sequence.any parameter to match current Swift stdlib.

Extend Bundle to provide URL scheme info in a nicer format.