Avoid flash of old/missing content after dropping loading view from in front of render view.
Moved to the WKNavigationDelegate's didFinish method, which seems to be the same as document.readyState == "complete". Not sure whether it's the right tradeoff, as now we obscure loaded content while waiting for other resources to load on the page, but it's better than flashing an empty screen so it's worth a shot.