Fix posts text jumping around when loading tweets.
The web view would quickly zoom out then back in when tweets show up. Since we don't have any need for people to zoom out, we can fix it by setting a minimum zoom scale of 1.

The `width=320` part is a minimum width; the web view will make the viewport wider to fill the available space. (Weird how convenient it is to actually read up on how the viewport meta tag works!)

Fixes #879.