Fix crash on orientation change.
Starting in iOS 13, Awful would sometimes crash immediately when changing the orientation of the device. TestFlight crash feedback pointed to an endless recursive call in `-targetDisplayModeForActionInSplitViewController:` which was indirectly attempting to access the split view controller's `displayModeButtonItem`, which would cause it to ask its delegate for the `-targetDisplayModeForActionInSplitViewController:`. Awful's implementation of this method was admittedly not what the method is made for, so let's try something a bit more proper.

Fixes #968.